Merged PR 226645: Filter out non-agent docker containers from environment
The DockerEnvironment currently lists all containers on the host. It should only list containers managed by the Agent. In order to achieve this, each container is labeled with an `"owner"` label set to `"Microsoft.Azure.Devices.Edge.Agent"` to designate that the agent created the container.
The environment then filters based on this label.

This PR also updates the test config to allow for updating of parameters per test environment. A `base.json` file is used for all common config. A test environment config (named as `<env>.json`) is loaded after the base to allow the parameters to be overridden. Test environment is set via `environment.json` or setting the `testEnvironment` env variable.

Note: the config changes allows the docker tests to be re-enabled on the build.