Merged PR 229675: Wire up autofac in the agent
This sets up Autofac (a dependency injection framework) to create the objects in the agent.

There were a couple of changes that were required:
* Introduced a `ISerde<T>` type
* Made all of the properties in `ConfigFileSource` `readonly`

All objects in `main` are now instantiated through the autofac container. A `Standalone` autofac module was created to set up the agent for running locally in dev mode. We can create another module when we are configuring via the twin.