Merged PR 265111: Start Docker containers in a way that actually catches signals.
Docker CMD statements in "shell form" execute the statement via `/bin/sh -c`. This shell catches the signals docker sends, and we don't see it. If we switch to the "exec form", the command is run as process 1, and docker will send the signals to our process.

Related work items: #2