Merged PR 272735: Set message properties on the mqtt topic for incoming and outgoing messages
Issue -
Currently, we were not handling messages with properties properly. The properties would not get set (or the message would fail to get processed properly).

Fix -
This PR retrieves the properties from the MQTT topic and sets it on the message for incoming messages, and does the reverse for the outgoing messages.

Related work items: #1270208