Merged PR 283406: Start edge hub first in edge-service
We want the agent to connect to IoT Hub through the edge hub, so the hub should start first. Note that both agent and hub are highly asynchronous, so putting hub first doesn’t guarantee it will run first, it just improves the odds. We’ll want to do more work to guarantee the hub is up when the agent tries to connect, especially if we see problems in this area.