Merged PR 282432: Make "journald" the default docker logger on linux.
For the private preview, we plan to use journald for docker logging.
This meets some of our P0 and P1 requirements:
- Creates logs in a standardized format
- Follows the common patterns and practices of docker containers - typically logging to STDOUT and STDERR.
- Will be organized by module.

Users can get all docker logs by running:
`journalctl --unit=docker`
Or get logs per module by running:
`journalctl -a CONTAINER_NAME=<module-name>`

Journald is not available on Windows, so I put this as a configuration in `appsettings.json`, so that we can still run on our dev boxes, and is overridden in the Docker files for Linux.