Merged PR 304737: Enable twin config source to retrieve config from backup
These changes enable the MMA to boot up successfully in the absence of connectivity to IoTHub. The MMA uses a backup copy of the module set in order to complete its boot up sequence. During normal (online, with connectivity to IoTHub) operation, the MMA updates the backup copy each time it receives a desired property update callback from IoTHub.

Please note that in the absence of connectivity to IoTHub, the MMA's request to register a desired property update callback will not be successful. The MMA does not currently retry its call to register a callback. As a result, the MMA will never successfully register a callback until it is restarted and has connectivity to IoTHub when booting up. Filed https://msazure.visualstudio.com/One/_workitems/edit/1338778 for this issue.