Merged PR 338414: Synchronize applying a patch with getting a twin
This change introduces synchronization between refreshing a twin (after connection status has toggled from offline to connected) and applying a patch. The reason we need this synchronization is so that we don't apply a patch to a stale copy of the twin - one that missed patch updates because the device was offline. As part of this change, the last known version of the twin is now cached in the TwinConfigSource and not the Agent.