Merged PR 433817: Local storage and offline availability of twin reported properties
These changes enable modules to update their twin's reported properties even when Edge Hub is offline. The TwinManager updates/merges the local copy of the twin's reported properties with the new patch of reported properties. It also collects all patches by merging them all in to a single patch so that when the cloud is back online, it can apply the collective patch to the online copy. This PR does not have changes to sync the online copy of the twin after connectivity is re-established. That will be in the next PR.