Merged PR 475834: Create a script to create multi-arch manifest
```
buildManifest.sh [options]
Note: Depending on the options you might have to run this as root or sudo.

options
 -r, --registry       Docker registry required to build, tag and run the module
 -u, --username       Docker Registry Username
 -p, --password       Docker Username's password
 -v, --image-version  Docker Image Version.
 -t, --template       Yaml file template for manifest definition.
```

This scripts assumes that the platform specific images are already in the docker registry. We need to use a yaml template because the underlying `manifest-tool` is not flexible enough to specify everything through command line args. I've included templates for all of our current images.