Merged PR 485678: Add script to build images one at a time
This script is based off of the `buildDockerImage.sh` script to allow images for single projects to be built. We need to be able to disable the building of the functions binding image for ARM for the time-being. Because all of the images to be built are hard-coded in the `buildDockerImage.sh`, turning this off requires a code change. This new script will allow us to pick which images to build in the build job without making a code change.

Once this is merged, we can update the build jobs, and potentially remove the `buildDockerImage.sh` script.

Script help (adds -n and -P options):
```
buildImage.sh [options]
Note: Depending on the options you might have to run this as root or sudo.

options
 -n, --image-name     Image name (e.g. edge-agent)
 -P, --project        Project to build image for (e.g. Microsoft.Azure.Devices.Edge.Agent.Service)
 -r, --registry       Docker registry required to build, tag and run the module
 -u, --username       Docker Registry Username
 -p, --password       Docker Username's password
 -v, --image-version  Docker Image Version. Either use this option or set env variable BUILD_BUILDNUMBER
 -t, --target-arch    Target architecture (default: uname -m)
--bin-dir             Directory containing the output binaries. Either use this option or set env variable BUILD_BINARIESDIRECTORY
```