Merged PR 490046: Upgrade to latest NuGet
I recently updated the private SDK repo with an FI from the public repo. This includes the changes that will allow us to append the user agent string with info about IoT Edge.

This change updates all the edge bits to use the latest NuGet packages for the device and service SDKs.