Merged PR 505048: Support private Docker Hub repositories
This PR changes the parsing of the image name to follow the docker cli convention. Images without a registry address default to `https://index.docker.io/v1/`. This string is used to look up credentials. The `iotedgectl` is updated to default to this address if no address is given when using `login`.