Merged PR 513451: Add moveImage script
Adds a moveImage script to allow moving images between docker registries.

A little background on why this is needed. I want to update the release process to publish our signed bits into a private ACR (EdgeRelease) and the test PyPi. This will let us test the new images and tool before putting into the public repos.

When we release to Docker Hub, we will simply copy all of the images from the private ACR to Docker Hub