Merged PR 516425: Report unknown status on Agent shutdown.
It is misleading to see the module status in the portal as "Running" when the entire edge runtime has been uninstalled on the device.
It was suggested to me to report the status of all modules currently deployed to Unknown, and report the deployment status as unknown as well.  This will be a clear indicator that, even if the other modules are running, while the Agent is down, all other status is not known.  This shutdown path is the same for "stop" as "uninstall", so the solution will work just as well on each.

Once Agent exits it main loop, it will make a final attempt to update the Twin reported status with "Unknown" for all modules, and Unknown for the deployment status.

Related work items: #1714134