Merged PR 530075: Added base dockerfiles for arm images.
We want edgehub and modules to run as less privileged users. That means we need to create a user in the dockerfile, and then set up to run as that user.

However, our build systems are all amd64 architecture, and the commands to create users will not run on the ARM docker images.

The simplest solution is to manually create base images that set up the new user, and any other commands we want to run, then the build systems can use these base images to copy the latest bits into.

In this pull request are the new docker files to set up the base image, and the modified docker files to use them.