Merged PR 554122: Regroup pull, create, update and start commands in health planner
For modules that get added or updated in a deployment, the `HealthRestartPlanner` groups the commands it prepares in the plan by Docker operation. All pulls go first, then the creates followed by the updates and then the starts. This doesn't quite reflect the fact that some of these commands have inter-dependencies. It doesn't make sense to run the _create_ command for a module if the _pull_ command failed. This PR re-organizes the commands to better reflect inter-command dependencies.

Now we create `GroupCommand` instances where each group contains a pull followed by a create and then optionally, a start (if the desired status is "running"). This way if the pull command fails then the create and the start don't get run.