Merged PR 555979: DeviceClient does not retry on EdgeHubConnectionException
Caught this running the long haul twin verification module. The DeviceClient does not retry on EdgeHubConnectionException and we throw that exception from EdgeHub (routing code, CloudEndpoint.cs) when the cloud proxy is not available. The cloud proxy is not available after the SAS token expires. What happens as a result is that the device client on the module side does not generate a new token and retry. Instead it passes the exception up to the module and the module dies. This is not good!

If we want the device client (on the module side or leaf side) to retry, it needs to be one of
typeof(IotHubClientTransientException),
            typeof(IotHubCommunicationException),
            typeof(ServerBusyException),
            typeof(IOException),
            typeof(TimeoutException),
            typeof(ObjectDisposedException),
            typeof(OperationCanceledException),
            typeof(TaskCanceledException),
            typeof(IotHubThrottledException),
            typeof(System.Net.Sockets.SocketException),

I think IOException is the most suitable, so made EdgeHubConnectionException inherit from IOException.