Merged PR 561214: Adding multiplexing support in EdgeHub
Overview - To enable multiplexing, Device SDK needs a way to refresh tokens for the existing connection links. We do that by dropping the client connection, which (if using the Device SDK) should automatically reconnect and pass in the new updated token. We use this token for the existing cloud connection. Because we don't want the existing cloud connection to die, we do this process a little bit before the token actually expires.
So this PR includes the following -
1. Added support to update tokens for an existing cloud connection (as explained above)
2. Created a type CloudConnection that manages cloud connections for a device/module (refactoring)
3. Updated ConnectionManager to have CloudConnections instead of CloudProxy objects
4. Updated Device SDK references to use the new device SDK. This was necessary because there were changes required to the Device SDK to support this.

Still working on manual and automated tests.