Merged PR 567077: Add transient retry when get twin call fails in EdgeAgentConnection
If the `GetTwinAsync` call in `EdgeAgentConnection` fails its possible that it might be due to transient network errors or because we are getting throttled by IoT Hub. If we didn't attempt a retry then this object would be stuck in an "error" state till either the connection status on the underlying device connection changes or we receive a patch deployment. So we implement an exponential back-off retry here which allows us to recover from this situation.