Merged PR 568049: Enable 'Open Test' context menu option in VS Test Explorer
It seems that double-clicking a test in VS's Test Explorer window wasn't taking us to the test's source code because of the following PropertyGroup element found in most of our CSPROJ files: `<DebugType>full</DebugType>`.

It may have originally been added to work around some issues in .NET Core (related to debugging, or using XUnit in VS...I'm not sure). After removing it, however, I'm still able to run the unit tests in Visual Studio. And now I can double click a test, or right click and select "Open test", or hit F12, to go to the source code.

It turns out we'll have to add this _back_ in order to enable code coverage in Visual Studio (as a workaround for [this](https://github.com/Microsoft/vstest/issues/800) issue). But we'll add it behind a flag and/or conditions so that you can get code coverage if you want it.