Merged PR 567635: Refactor identity factory
Identity factory API accepted an MQTT like username and password. Since it is a generic component, we should remove the MQTT ness of the API.
This PR does the refactoring to do that.
The MQTT portion of it has been moved to the MQTT assembly. Also refactored the tests as appropriate.
No change in functionality. No new resharper issues. :)