Merged PR 575976: Update EntityStore update APIs to return the updated value and some test updates/refactoring
1. Updated the EntityStore Update* APIs to return the updated value
2. Updated uses of these APIs to reflect the change
3. Refactored tests so that the same set of tests run against InMemoryStore and RocksDB
4. Added some tests

One place where we should update the use of the EntityStore APIs but I didn't is TwinManager - primarily because making the changes caused several tests to fail. I took a look and seems like the tests will need to be modified significantly. I will take another stab at it later.