Merged PR 589474: Add nuget.org to list of package sources in nuget.config
The common VSTS tool for "dotnet restore" can either be set up with a list of feeds, or to look for a nuget.config file in the project.  I chose to use the "nuget.config" file, but that means we need to add a nuget.org reference into the file for the VSTS restore tool to work properly.

Added nuget.org to list of package sources in nuget.config