Merged PR 594753: Automate 1st tutorial for use in service deployments
This change contains one xunit test, annotated with the `[Deploy]` attribute. It always installs `iotedgectl` from pip, and always pulls docker images from dockerhub, so it's useful to the platform service team when they want to make sure their service deployments don't break IoT Edge. But we'll need to paramterize this test a bit if we want to validate builds prior to release.