Merged PR 603969: Override defaults with environment variables in automated tutorial
The following customization points are now available when you run the automated tutorial:

- Use $iotedgectlArchivePath to specify the location of the python bootstrap package (default is pypi)
- Use $iothubConnectionString to specify the IoT Hub connection string (default comes from KeyVault)
- Use $registryAddress|User|Password to specify the container registry (default is DockerHub)
- Use $eventhubCompatibleEndpointWithEntityPath to specify Event Hub-compatible endpoint with Entity Path (default comes from KeyVault)

I added debug output at each of the above points to document which option is chosen (Note: Console.WriteLine() doesn’t print to stdout on Windows, but it seems to work on Linux, including VSTS).

Finally, I added the ability to replace the Docker image versions for the agent, hub, and temp sensor, but I’m not actually replacing them yet because it won’t be necessary until we exit public preview.

Related work items: #1768296