Merged PR 631828: Make the nightly build a CI build
This PR renames the old CI build to add a `Checkin` suffix and updates it to only be used as a checkin gate. The Nightly builds (Linux and Windows) are updated to run as "CI" builds and are triggered after a PR is completed. They no longer run at night. The trigger is coalesced so that merges that occur while the build is running will batch.

This allows us to get faster feedback that a change has broken the build.