Merged PR 628229: Add AMQP head with SAS authentication support
This PR should be considered as _in progress_ right now. While the SAS authentication is implementation complete, I haven't really tested it with a Device SDK client because there are other pieces that need to be in place before we can do that. The following files in this PR are there only to make stuff compile and isn't ready for review:

1.   AmqpConnectionGatewayContext.cs
2.   AmqpRuntimeProvider.cs

Begin the review by reading `/doc/AmqpBootstrapImplementationNotes.md`. That should give you an idea about how AMQP SASL authentication works. `AmqpProtocolHead.cs` is the entrypoint - this object's `StartAsync` method is invoked when the Edge Hub service starts. `EdgeHubSaslPlainAuthenticator.cs` contains the code that implements SAS authentication.