Merged PR 645499: Setting certificate version to v3 to match the extensions. Yes the value
Setting certificate version to v3 so as to match the v3 extensions that are currently being added. Formerly no explicit version was being set and it was defaulting to v1. Yes it the decimal value 2 sets the version to 3. Sigh... openssl...

What this fixes is Java based certificate tooling which did not like the fact that v3 extensions were added to the certificate but the version was not v3.

Verification steps:
1) Validated certificate chain using openssl s_server and s_client.
2) Validated that Java's keytool imports the Edge server certificate, prior to this fix the error observed was "not a x.509 certificate"
3) Validated E2E Edge as a gateway scenario.